backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.4.0
psycopg[binary]==3.2.3
stripe==7.8.0
python-decouple==3.8
Pillow==10.4.0
django-extensions==3.2.3
requests==2.32.3
celery==5.4.0
redis==5.1.1
whitenoise==6.8.2
gunicorn==23.0.0
